Separated alert command Updated methods removeBroadcastWithType and removeLatestBroadcastWithType Removed unnecessary import
