+++ eternalcore-core/src/main/java/com/eternalcode/core/translation/Translation.java
@@ -6,7 +6,6 @@ import com.eternalcode.core.feature.language.Language;
 import com.eternalcode.multification.notice.Notice;
 import org.bukkit.Material;
 import org.bukkit.event.entity.EntityDamageEvent;
-import org.checkerframework.checker.units.qual.N;
 
 import java.util.Collection;
 import java.util.HashMap;
